Rename OthersPanel to ParticipantsPanel and Update ParticipantsPanel
