+++ src/com/thunder_cut/grapics/ui/ToolPanel.java
@@ -18,7 +18,6 @@ public class ToolPanel {
         //toolPanel.setLayout(new GridLayout(1, 4, 10, 10));
         toolPanel.setPreferredSize(new Dimension(1280, 180));
         toolPanel.setBackground(Color.LIGHT_GRAY);
-        toolPanel.setVisible(true);
 
         toolPanel.add(brush);
         toolPanel.add(eraser);
